backend/requirements.txt
gunicorn==20.1.0
whitenoise==6.4.0
Django==4.2.23
django-createsuperuser==2020.12.3
djangorestframework==3.14.0
django-filter===24.3
cryptography>=42.0
django-cors-headers~=3.1.1
drf-jwt~=1.19.2
pyjwt>=2.8.0
requests===2.32.5
psycopg==3.2.9
python-dateutil===2.9.0
urllib3===2.5.0
python-docx===1.2.0
webvtt-py===0.5.1
yt-dlp===2025.9.26